vulkaninfo: make where output goes uniform
Currently, text and json output are put to standard output while
html is put to vulkaninfo.html. This commit makes it uniform so that
unless specifed, all output types are put to standard out. By using
the `--dest=<filename>` command line argument a user can manually
specify the file name of where to put the output.
In this implementation, the file is appended with the extension
appropriate for the output, so .txt for the default text, .html for
html, and .json for json.

This is a breaking change in that --html will not create a new file,
but this brings the application back in line with user expectations.

files modified
 - vulkaninfo/vulkaninfo.cpp
 - vulkaninfo/vulkaninfo.md

Change-Id: I97004a1743e13f938b4b0210ad8373abc5e57dd3